+++ android/src/main/java/com/revenuecat/purchases_flutter/PurchasesFlutterPlugin.java
@@ -52,7 +52,7 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
     @Nullable private Activity activity;
 
     private static final String PLATFORM_NAME = "flutter";
-    private static final String PLUGIN_VERSION = "3.5.0";
+    private static final String PLUGIN_VERSION = "3.8.0";
 
     /**
      * Plugin registration.
@@ -210,6 +210,9 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
             case "isAnonymous":
                 isAnonymous(result);
                 break;
+            case "isConfigured":
+                isConfigured(result);
+                break;
             case "checkTrialOrIntroductoryPriceEligibility":
                 productIdentifiers = call.argument("productIdentifiers");
                 checkTrialOrIntroductoryPriceEligibility(productIdentifiers, result);
@@ -262,6 +265,10 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
             case "setOnesignalID":
                 String onesignalID = call.argument("onesignalID");
                 setOnesignalID(onesignalID, result);
+                break;
+            case "setAirshipChannelID":
+                String airshipChannelID = call.argument("airshipChannelID");
+                setAirshipChannelID(airshipChannelID, result);
                 break;
             case "setMediaSource":
                 String mediaSource = call.argument("mediaSource");
@@ -460,6 +467,10 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
         result.success(CommonKt.isAnonymous());
     }
 
+    private void isConfigured(final Result result) {
+        result.success(Purchases.isConfigured());
+    }
+
     private void checkTrialOrIntroductoryPriceEligibility(ArrayList<String> productIDs, final Result result) {
         result.success(CommonKt.checkTrialOrIntroductoryPriceEligibility(productIDs));
     }
@@ -523,7 +534,12 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
         result.success(null);
     }
 
-    private void setMediaSource(String mediaSource, final Result result) { 
+    private void setAirshipChannelID(String airshipChannelID, final Result result) {
+        SubscriberAttributesKt.setAirshipChannelID(airshipChannelID);
+        result.success(null);
+    }
+
+    private void setMediaSource(String mediaSource, final Result result) {
         SubscriberAttributesKt.setMediaSource(mediaSource);
         result.success(null);
     }